csf functions returns integer

+++ src/QCtools.jl
@@ -111,7 +111,7 @@ function csf(n, m, S)
   a = (2*S+1)/(m+1)
   b = binomial(m+1,Int(n/2) - S)
   c = binomial(m+1,Int(n/2) + S +1)
-  return a*b*c
+  return Int(a*b*c)
 end
 
 end
